Split CreateNewsPositiveTest to Positive, Negative and Preview test classes
